Worked out all compiler warnings. No memory leaks reported by Valgrind.

+++ parser/wgsl-parser.c
@@ -54,21 +54,21 @@ int parse_function_decl(char *shader_code, int *workgroup_sizes,
 
     struct slre_cap caps[4]; // Up to three numbers and the function name
     int match_length;
-    for (unsigned long i = 0; i < sizeof(patterns) / sizeof(patterns[0]); i++) {
-        match_length = slre_match(patterns[i], shader_code, strlen(shader_code),
-                                  caps, 4, 0);
+    for (int i = 0; i < ((int)(sizeof(patterns) / sizeof(patterns[0]))); i++) {
+        match_length = slre_match(patterns[i], shader_code,
+                                  (int)strlen(shader_code), caps, 4, 0);
         if (match_length > 0) { // Match found
             // Extract the function name from the last capture group
             int fn_index = i + 1; // Function name index depends on the number
                                   // of workgroup sizes
-            memcpy(function_name, caps[fn_index].ptr, caps[fn_index].len);
+            memcpy(function_name, caps[fn_index].ptr, (size_t)caps[fn_index].len);
             function_name[caps[fn_index].len] = '\0';
 
             // Extract workgroup sizes
-            for (unsigned long j = 0; j <= i; j++) {
+            for (int j = 0; j <= i; j++) {
                 if (caps[j].len > 0) {
                     char size_str[10];
-                    memcpy(size_str, caps[j].ptr, caps[j].len);
+                    memcpy(size_str, caps[j].ptr, (size_t)caps[j].len);
                     size_str[caps[j].len] = '\0';
                     workgroup_sizes[j] = atoi(size_str);
                 }
@@ -94,7 +94,8 @@ int parse_binding_defs(char *shader_code, ComputeInfo *info) {
         printf("Error -> parse_binding_defs(): ComputeInfo is NULL\n");
         return -1;
     }
-
+    
+    // Save the start of the shader code for resetting the pointer
     char *shader_code_start = shader_code;
     int match_length;
     int num_found = 0;
@@ -106,18 +107,19 @@ int parse_binding_defs(char *shader_code, ComputeInfo *info) {
 
     // Iterate until all patterns are tested.
     // If no pattern is found, return -1
-    for (unsigned long i = 0; i < sizeof(patterns) / sizeof(patterns[0]); i++) {
+    for (int i = 0; i < ((int)(sizeof(patterns) / sizeof(patterns[0]))); i++) {
         // (group) (binding) (type)
         struct slre_cap caps[3];
         // Iterate through the shader code and find all matches of the pattern
-        while ((match_length = slre_match(patterns[i], shader_code,
-                                          strlen(shader_code), caps, 3, 0))) {
+        while (
+            (match_length = slre_match(patterns[i], shader_code,
+                                       (int)strlen(shader_code), caps, 3, 0))) {
             if (match_length > 0 &&
                 num_found < MAX_GROUPS) { // Check bounds for array
                 int group = atoi(caps[0].ptr);
                 int binding = atoi(caps[1].ptr);
                 strncpy(info->groups[group].bindings[binding].type, caps[2].ptr,
-                        caps[2].len);
+                        (size_t)caps[2].len);
                 info->groups[group].bindings[binding].type[caps[2].len] =
                     '\0'; // Null-terminate the string
                 num_found++;
